+++ cpp/include/cudf/types.h
@@ -4,10 +4,11 @@
 typedef int gdf_size_type; ///< Limits the maximum size of a gdf_column to 2^31-1
 typedef gdf_size_type gdf_index_type;
 typedef unsigned char gdf_valid_type;
-typedef  long  gdf_date64;
-typedef  int    gdf_date32;
-typedef  int    gdf_category;
-typedef  long  gdf_timestamp;
+typedef	long	gdf_date64;
+typedef	int		gdf_date32;
+typedef	int		gdf_category;
+typedef	long	gdf_timestamp;
+typedef signed char gdf_bool;
 typedef int   gdf_nvstring_category;
 
 
@@ -27,6 +28,7 @@ typedef enum {
     GDF_TIMESTAMP,  ///< Exact timestamp encoded with int64 since UNIX epoch (Default unit millisecond)
     GDF_CATEGORY,
     GDF_STRING,
+    GDF_BOOL,
     GDF_STRING_CATEGORY, ///< Stores indices of an NVCategory in data and in extra col info a reference to the nv_category
     N_GDF_TYPES,   ///< additional types should go BEFORE N_GDF_TYPES
 } gdf_dtype;