+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # cuDF 0.7.0 (Date TBD)
 
 ## New Features
+- PR #1142 Added `GDF_BOOL` column type
 
 ...
 
@@ -20,6 +21,7 @@
 
 - PR #760 Raise `FileNotFoundError` instead of `GDF_FILE_ERROR` in `read_csv` if the file does not exist
 - PR #539 Add Python bindings for replace function
+- PR #817 Added GDF_BOOL data type to CUDA code
 - PR #823 Add Doxygen configuration to enable building HTML documentation for libcudf C/C++ API
 - PR #807 CSV Reader: Add byte_range parameter to specify the range in the input file to be read
 - PR #857 Add Tail method for Series/DataFrame and update Head method to use iloc
@@ -65,6 +67,7 @@
 - PR #1159 Add Python bindings for libcudf dlpack functions
 - PR #1155 Add __array_ufunc__ for DataFrame and Series for sqrt
 - PR #1168 to_frame for series accepts a name argument
+
 
 ## Improvements